Sync zsh_history per machine
The Problem:
============
I'd like a more thorough overview of my CLI history in order to look for
optimizations, patterns and other such things. However, with each
machine only storing it's history locally and with a cap of 1k lines,
I'm losing it all.

The Solution:
=============
Increase history to 50k lines. We'll see if that slows startup.

Move history to a synced folder and a file named after the current host;
this stops the different machines generating conflicts and my work-day
history from polluting my personal history.